笔记大纲初始化

+++ doc/server-side/GraphQL + mongoDB/server2/src/graphql/typeDefs.ts
@@ -1,10 +1,4 @@
-/**
- * GraphQL的AST抽象语法，简化开发
- *
- *
- */
-
-import { gql } from "apollo-server-express";
+import { gql } from 'apollo-server-express';
 
 export const typeDefs = gql`
   type Listing {
@@ -16,11 +10,13 @@ export const typeDefs = gql`
     numOfGuests: Int!
     numOfBeds: Int!
     numOfBaths: Int!
-    rating: Int!
+    rating: Float!
   }
+
   type Query {
     listings: [Listing!]!
   }
+
   type Mutation {
     deleteListing(id: ID!): Listing!
   }
